config service: split into service and controller
